UsernameService fix 4

+++ src/test/resources/init_dish_order.sql
@@ -4,6 +4,6 @@ CREATE TABLE Dishes_Orders
   id       BIGINT NOT NULL AUTO_INCREMENT PRIMARY KEY,
   order_id BIGINT NOT NULL,
   dish_id  BIGINT NOT NULL,
-  amount   TINYINT(4)  NOT NULL,
-  item_sum INT(11)     NOT NULL
+  amount   BIGINT(4)  NOT NULL,
+  item_sum BIGINT(11)     NOT NULL
 );
